Multiple games rendered, breaks after first guess
Error says guess is already full on the first guess even after pressing
enter

Seems like enter is not actually registering... needs more research
