+++ package.json
@@ -15,6 +15,7 @@
         "random-words": "^2.0.0",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
+        "react-promise-tracker": "^2.1.1",
         "react-scripts": "5.0.1",
         "typescript": "^4.9.5",
         "uuid": "^9.0.1",
@@ -51,4 +52,3 @@
         "tailwindcss": "^3.4.0"
     }
 }
-
