+++ src/components/board/Board.tsx
@@ -1,32 +1,14 @@
 import { v4 as uuidv4 } from 'uuid';
 import Guess from '../../types/Guess';
 import GuessRow from './GuessRow';
-import { useEffect, useState } from 'react';
 
 export interface BoardProps {
     guesses: Guess[];
     word: string;
+    isWord: boolean[];
 }
 
-const Board = ({ guesses, word }: BoardProps) => {
-    const [isWord, setIsWord] = useState<boolean[]>([]);
-
-    useEffect(() => {
-        const updateIsWord = async () => {
-            const newIsWord: boolean[] = Array(guesses.length);
-
-            for (let i = 0; i < newIsWord.length; i++) {
-                if (!guesses[i].isFilled || (await !guesses[i].isWord())) {
-                    newIsWord[i] = true;
-                }
-            }
-
-            setIsWord(newIsWord);
-        };
-
-        updateIsWord();
-    }, [guesses]);
-
+const Board = ({ guesses, word, isWord }: BoardProps) => {
     return (
         <div className={`grid gap-1 w-1/3 h-full m-auto overflow-y-auto`}>
             {guesses.map((guess: Guess, index: number) => (
