+++ src/index.tsx
@@ -1,9 +1,9 @@
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import Game from './Game';
+import App from './App';
 
 const root = ReactDOM.createRoot(
     document.getElementById('root') as HTMLElement,
 );
-root.render(<Game />);
+root.render(<App />);
